# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-vendor e-commerce API for WhatsApp sellers
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariant } from "./productVariant";

export interface CreateProductBody {
  name: string;
  description?: string | null;
  price: number;
  categoryId?: number | null;
  images: string[];
  variants: ProductVariant;
  inStock: boolean;
  featured: boolean;
}