# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-vendor e-commerce API for WhatsApp sellers
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}