# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-vendor e-commerce API for WhatsApp sellers
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  storeId: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  notes?: string | null;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  whatsappMessage: string;
  createdAt: string;
}